pyproject.toml
[project]
name = "study-adkmemory-cpuse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.4.1",
    "google-cloud-aiplatform==1.97.0",
    "google-genai==1.20.0",
    "python-dotenv>=1.1.1",
]
